# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Confecciones WA",
        "description": "Documentación de la API para el sistema de Confecciones WA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Listar facturas",
                "description": "Obtiene el listado paginado de facturas.",
                "operationId": "0757acd1a0a3d8632bb68443f2c8cda3",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por número de factura.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "INV-001"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de registros por página.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10,
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facturas obtenidas correctamente"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Registrar factura",
                "description": "Crea una nueva factura.",
                "operationId": "7a85316a950ac312f5bd5ce684371e00",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "invoice_number",
                                    "total_amount",
                                    "issue_date",
                                    "is_active"
                                ],
                                "properties": {
                                    "production_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "invoice_number": {
                                        "type": "string",
                                        "example": "INV-001"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "currency": {
                                        "description": "USD o PEN",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-30"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-30",
                                        "nullable": true
                                    },
                                    "payment_status": {
                                        "description": "PENDIENTE, PAGADA, ANULADA",
                                        "type": "string",
                                        "example": "PENDIENTE"
                                    },
                                    "attached_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Factura creada correctamente"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/invoices/{id}": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "Actualizar factura",
                "description": "Actualiza la información de una factura. Usa POST enviando _method=PUT debido al soporte de envío de archivos.",
                "operationId": "a7b3f28c56fb6181037bd6a29276fe72",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la factura",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "invoice_number",
                                    "total_amount",
                                    "issue_date",
                                    "is_active",
                                    "_method"
                                ],
                                "properties": {
                                    "production_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    },
                                    "invoice_number": {
                                        "type": "string",
                                        "example": "INV-001"
                                    },
                                    "total_amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1500.5
                                    },
                                    "currency": {
                                        "description": "USD o PEN",
                                        "type": "string",
                                        "example": "USD"
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-07-30"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2026-08-30",
                                        "nullable": true
                                    },
                                    "payment_status": {
                                        "description": "PENDIENTE, PAGADA, ANULADA",
                                        "type": "string",
                                        "example": "PENDIENTE"
                                    },
                                    "attached_file": {
                                        "type": "string",
                                        "format": "binary",
                                        "nullable": true
                                    },
                                    "_method": {
                                        "description": "Requerido para simular PUT via POST",
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "is_active": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Factura actualizada correctamente"
                    },
                    "404": {
                        "description": "Factura no encontrada"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}